+++ spec/spec_helper.rb
@@ -185,6 +185,10 @@ module Project
       def update?
         true
       end
+
+      def destroy?
+        false
+      end
     end
   end
 end